fix: correct defaultNumCtx type (int, not uint64)
The defaultNumCtx field is typed as int but the TurboQuant context
scaling was assigning a uint64 value, causing a compilation error
on Windows (GOOS=windows).

+++ server/routes.go
@@ -1878,7 +1878,7 @@ func Serve(ln net.Listener) error {
 			if contextMultiplier < 1.0 {
 				contextMultiplier = 1.0
 			}
-			newCtx := uint64(float64(s.defaultNumCtx) * contextMultiplier)
+			newCtx := int(float64(s.defaultNumCtx) * contextMultiplier)
 			slog.Info("turboquant context scaling",
 				"original_ctx", s.defaultNumCtx,
 				"compression_ratio", fmt.Sprintf("%.1fx", ratio),
